CRT & nCr

+++ 6. Geometry/Basic Geo.cpp
@@ -8,7 +8,7 @@ struct pt {
 
 /* https://cp-algorithms.com/geometry/oriented-triangle-area.html
    ccw = +ve, cw = -ve, linear = 0 */
-int orient(pt a, pt b, pt c)
+double orient(pt a, pt b, pt c)
 {
     return a.x * (b.y-c.y) + b.x * (c.y-a.y) + c.x * (a.y-b.y);
 }
